+++ File/ResolvedReference/ResolvedRequestFileReference.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Core\File\ResolvedReference;
 
 use Heptacom\HeptaConnect\Core\Bridge\File\FileRequestUrlProviderInterface;
-use Heptacom\HeptaConnect\Core\Storage\RequestStorage;
+use Heptacom\HeptaConnect\Core\Storage\Contract\RequestStorageContract;
 use Heptacom\HeptaConnect\Portal\Base\File\ResolvedFileReferenceContract;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\FileReferenceRequestKeyInterface;
@@ -19,14 +19,14 @@ class ResolvedRequestFileReference extends ResolvedFileReferenceContract
 
     private FileRequestUrlProviderInterface $fileRequestUrlProvider;
 
-    private RequestStorage $requestStorage;
+    private RequestStorageContract $requestStorage;
 
     public function __construct(
         PortalNodeKeyInterface $portalNodeKey,
         FileReferenceRequestKeyInterface $requestId,
         ClientInterface $client,
         FileRequestUrlProviderInterface $fileRequestUrlProvider,
-        RequestStorage $requestStorage
+        RequestStorageContract $requestStorage
     ) {
         parent::__construct($portalNodeKey);
         $this->requestId = $requestId;